pyproject.toml
[project]
name="synthesizer"
version="0.0.1"
dependencies=[
    "cohere==5.6.2",
    "ollama",
    "python-dotenv==1.0.1",
    "langchain_community==0.2.11",
    "langchain_text_splitters==0.2.2",
    "pandas==2.2.2",
    "numpy==1.26.4",
    "kneed==0.8.5",
    "jinja2==3.1.4",
    "random11==0.0.1",
    "loguru==0.7.2",
    "scikit-learn==1.5.1",
    "tiktoken==0.3.3", 
    "pypdf==4.3.1", 
    "kneed[plot]==0.8.5", 
    "openpyxl==3.1.5",  
    "arize-phoenix==4.33.1", 
    "langchain_cohere==0.2.4", 
    "opentelemetry-sdk==1.27.0", 
    "opentelemetry-exporter-otlp==1.27.0",
    "langchain-ollama==0.1.3"
]
[tool.setuptools]
packages=["synthesizer"]
